LOK: make getViews() be a member function of Document
Just to be consistent, as all other view-related member functions are
there, too.

No real impact, as only the unit test uses this so far, and it always
works with a single document.

(cherry picked from commit a04b31c9facb08a8e38b79480b9a73bd2693cb9e)

+++ include/sfx2/lokhelper.hxx
@@ -25,9 +25,8 @@ public:
     static void setView(size_t nId);
     /// Get the currently active view.
     static size_t getView();
-
-    /// Total number of view shells.
-    static int getViews();
+    /// Get the number of views of the current object shell.
+    static size_t getViews();
 };
 
 #endif
